separate BoolCompare and LazyOr to separate packages to not conflict
